small improvs
Fixed issue that in some rare cases swap was not removed and need to retry.
Default and tools site cannot be deleted, now can be forced with delete-all or delete=force and is better handled to automaticalle set to default when forced.
Fixed issue with cloned sites replacing content URL wrong in some cases.